% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{multi_plot}
\alias{multi_plot}
\title{Plots exposures and frequency by given variables}
\usage{
multi_plot(dat, vars, claim_var = "ClaimNb", exposure_var = "Exposure")
}
\arguments{
\item{dat}{input data set}

\item{vars}{vector of variable names in character format i.e
\verb{c("var1, "var2")}}

\item{claim_var}{column that represents the number of claims for an obs}

\item{exposure_var}{column that represents the number of exposures for an obs}
}
\value{
ggplots of exposures and frequency, grouped by the variables in vars
}
\description{
Plots exposures and frequency by given variables
}
