% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{prep_lgb}
\alias{prep_lgb}
\title{Prep for LightGBM training}
\usage{
prep_lgb(dat, x)
}
\arguments{
\item{dat}{dataframe}

\item{x}{vector of covariates i.e \code{c("DrivAge", "VehAge")}}
}
\value{
a data matrix
}
\description{
Creates the data matrix needed to construct a LightGBM dataset from the given
covariates 'x'
}
