% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_search_par.R
\name{random_search_par}
\alias{random_search_par}
\title{Parallelized random search}
\usage{
random_search_par(bounds, niter, training_data, n_cores = NULL)
}
\arguments{
\item{bounds}{Lower and Upper bounds for selected hyperparameters. See
\code{scoring_function()} docs for params to train}

\item{niter}{Number of random search iterations to perform}

\item{training_data}{Training data for LightGBM model}

\item{n_cores}{How many cores to split search across}
}
\value{
DataFrame of results
}
\description{
Parallelized random search
}
